// src/lib/openapi/squad/models/CreateRequirement.ts
export interface CreateRequirement {
    /**
     * 
     * @type {string}
     * @memberof CreateRequirement
     */
    id?: string;
    /**
     * System prompt used to generate the requirement
     * @type {string}
     * @memberof CreateRequirement
     */
    systemPrompt?: string | null;
    /**
     * Current state of AI processing
     * @type {string}
     * @memberof CreateRequirement
     */
    aiProcessingState: CreateRequirementAiProcessingStateEnum;
    /**
     * Log of refinements made to the requirement
     * @type {Array<SolutionRefinementLogInner>}
     * @memberof CreateRequirement
     */
    refinementLog: Array<SolutionRefinementLogInner>;
    /**
     * Title of the requirement
     * @type {string}
     * @memberof CreateRequirement
     */
    title: string;
    /**
     * Description of the requirement
     * @type {string}
     * @memberof CreateRequirement
     */
    description: string;
    /**
     * The feature this requirement belongs to
     * @type {string}
     * @memberof CreateRequirement
     */
    feature: string;
    /**
     * Detailed requirements specification
     * @type {string}
     * @memberof CreateRequirement
     */
    requirements: string;
    /**
     * ID of the requirement owner
     * @type {string}
     * @memberof CreateRequirement
     */
    ownerId?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateRequirement
     */
    status: CreateRequirementStatusEnum;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateRequirement
     */
    solutionIds?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateRequirement
     */
    opportunityIds?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateRequirement
     */
    feedbackIds?: Array<string>;
}


/**
 * @export
 */
export const CreateRequirementAiProcessingStateEnum = {
    Initial: 'INITIAL',
    Processing: 'PROCESSING',
    Finished: 'FINISHED',
    Error: 'ERROR'
} as const;
export type CreateRequirementAiProcessingStateEnum = typeof CreateRequirementAiProcessingStateEnum[keyof typeof CreateRequirementAiProcessingStateEnum];

/**
 * @export
 */
export const CreateRequirementStatusEnum = {
    Draft: 'draft',
    Ready: 'ready',
    Complete: 'complete'
} as const;
export type CreateRequirementStatusEnum = typeof CreateRequirementStatusEnum[keyof typeof CreateRequirementStatusEnum];